integrate lock approval revoke into claim airdrop flow
