+++ src/components/veocean/VeOceanPortal.svelte
@@ -2,27 +2,11 @@
   import { userAddress, connectedChainId } from "../../stores/web3";
   import OceanSummary from "./OceanSummary.svelte";
   import VeOceanCard from "./VeOceanCard.svelte";
-  import Button from "../common/Button.svelte";
   import { getLockedOceanAmount, getLockedEndTime } from "../../utils/ve";
   import { lockedOceanAmount, oceanUnlockDate } from "../../stores/veOcean";
-  import { ToastNotification } from "carbon-components-svelte";
-  import {
-    approve as approveToken,
-    allowance
-  } from "../../utils/tokens";
-  import { getAddressByChainIdKey } from "../../utils/address/address";
   import moment from "moment";
-  import { ethers } from "ethers";
 
   let loading = false;
-  let showDismissAllowance = false;
-  let allowedTokenAmt = 0;
-  const supportedChainId = import.meta.env.VITE_VE_SUPPORTED_CHAINID;
-
-  const setShowApprovalNotification = async(value, allowedTokens) => {
-    allowedTokenAmt=allowedTokens
-    showDismissAllowance=value
-  }
 
   const loadValues = async () => {
     loading = true;
@@ -47,43 +31,8 @@
     }
   })
 
-  const dismissTokenApproval = async() =>{
-    try{
-      await approveToken(
-      getAddressByChainIdKey($connectedChainId, "Ocean"),
-      getAddressByChainIdKey(
-        supportedChainId,
-        "veOCEAN"
-      ),
-      0
-      )
-      setShowApprovalNotification(false, 0)
-    }catch(e){
-      console.error(e)
-    }
-
-  }
-
   $: if ($userAddress && $connectedChainId==import.meta.env.VITE_VE_SUPPORTED_CHAINID) {
     loadValues();
-    allowance(
-      getAddressByChainIdKey($connectedChainId, "Ocean"),
-      $userAddress,
-      getAddressByChainIdKey(
-       supportedChainId,
-       "veOCEAN"
-     )
-    ).then((allowedAmt) => {
-      if(allowedAmt>0){
-        allowedTokenAmt = ethers.utils.formatEther(
-          BigInt(allowedAmt).toString(10)
-        )
-        showDismissAllowance = true
-      }else{
-        allowedTokenAmt = 0
-        showDismissAllowance = false
-      }
-    })
   }
 </script>
 
@@ -94,27 +43,6 @@
   </h2>
   </div>
   <div class={`container`}>
-  {#if showDismissAllowance}
-    <ToastNotification
-      fullWidth
-      lowContrast
-      hideCloseButton
-      kind="warning"
-      title={`REVOKE THE ${allowedTokenAmt > 1000000 ? '>1000000.00' : parseFloat(allowedTokenAmt).toFixed(2)} TOKEN LOCK APPROVAL!!`}
-      subtitle="Passive-DF is stoped. Remove the current token approval to make sure no one else could lock the approved token in your behalf!"
-      on:close={(e) => {
-        e.preventDefault();
-        showDismissAllowance = false;
-      }}
-    >
-      <Button
-        className="dismissAllowanceButton"
-        text={"revoke token approval"}
-        onclick={dismissTokenApproval}
-        disabled={allowedTokenAmt==0}
-      />
-    </ToastNotification>
-  {/if}
     <VeOceanCard />
     <OceanSummary />
   </div>
