+++ src/utils/metadata/epochs/epochs.json
@@ -2766,21 +2766,21 @@
     "streams": [
       {
         "name": "Passive",
-        "rewards": 150000,
+        "rewards": 0,
         "substreams": [
           {
             "name": "veOCEAN",
-            "rewards": 150000
+            "rewards": 0
           }
         ]
       },
       {
         "name": "Active",
-        "rewards": 150000,
+        "rewards": 37500,
         "substreams": [
           {
             "name": "Volume DF",
-            "rewards": 112500
+            "rewards": 0
           },
           {
             "name": "Predictoor DF",
@@ -2789,7 +2789,7 @@
         ]
       }
     ],
-    "total": 300000
+    "total": 37500
   },{
     "id": 89,
     "date_start": "2024-05-09",
@@ -2797,21 +2797,21 @@
     "streams": [
       {
         "name": "Passive",
-        "rewards": 150000,
+        "rewards": 0,
         "substreams": [
           {
             "name": "veOCEAN",
-            "rewards": 150000
+            "rewards": 0
           }
         ]
       },
       {
         "name": "Active",
-        "rewards": 150000,
+        "rewards": 37500,
         "substreams": [
           {
             "name": "Volume DF",
-            "rewards": 112500
+            "rewards": 0
           },
           {
             "name": "Predictoor DF",
@@ -2820,7 +2820,7 @@
         ]
       }
     ],
-    "total": 300000
+    "total": 37500
   },{
     "id": 90,
     "date_start": "2024-05-16",
@@ -2828,21 +2828,21 @@
     "streams": [
       {
         "name": "Passive",
-        "rewards": 150000,
+        "rewards": 0,
         "substreams": [
           {
             "name": "veOCEAN",
-            "rewards": 150000
+            "rewards": 0
           }
         ]
       },
       {
         "name": "Active",
-        "rewards": 150000,
+        "rewards": 37500,
         "substreams": [
           {
             "name": "Volume DF",
-            "rewards": 112500
+            "rewards": 0
           },
           {
             "name": "Predictoor DF",
@@ -2851,6 +2851,6 @@
         ]
       }
     ],
-    "total": 300000
+    "total": 37500
   }
 ]
